# Cache property mapping in DerivedTypeConverter<T>